iast tests, separate file and upgrade
