+++ tests/contrib/django/test_django_appsec.py
@@ -2,16 +2,11 @@
 import json
 import logging
 
-import mock
 import pytest
 
 from ddtrace import config
-from ddtrace._monkey import patch_iast
 from ddtrace.appsec._constants import APPSEC
-from ddtrace.appsec._constants import IAST
 from ddtrace.appsec._constants import SPAN_DATA_NAMES
-from ddtrace.appsec.iast import oce
-from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
 from ddtrace.ext import http
 from ddtrace.internal import _context
 from ddtrace.internal import constants
@@ -397,18 +392,6 @@ def test_django_client_ip_header_set_by_env_var_invalid_2(client, test_spans, tr
         assert not root_span.get_tag(http.CLIENT_IP)
 
 
-def test_django_weak_hash(client, test_spans, tracer):
-    with override_global_config(dict(_appsec_enabled=True, _iast_enabled=True)):
-        oce.reconfigure()
-        patch_iast(weak_hash=True)
-        root_span, _ = _aux_appsec_get_root_span(client, test_spans, tracer, url="/appsec/weak-hash/")
-        str_json = root_span.get_tag(IAST.JSON)
-        assert str_json is not None, "no JSON tag in root span"
-        vulnerability = json.loads(str_json)["vulnerabilities"][0]
-        assert vulnerability["location"]["path"].endswith("tests/contrib/django/django_app/appsec_urls.py")
-        assert vulnerability["evidence"]["value"] == "md5"
-
-
 def test_request_ipblock_403(client, test_spans, tracer):
     """
     Most blocking tests are done in test_django_snapshots but
@@ -746,402 +729,3 @@ def test_request_suspicious_request_block_match_response_headers(client, test_sp
     with override_global_config(dict(_appsec_enabled=False)), override_env(dict(DD_APPSEC_RULES=RULES_SRB)):
         root_span, response = _aux_appsec_get_root_span(client, test_spans, tracer, url="/appsec/response-header/")
         assert response.status_code == 200
-
-
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)):
-        oce.reconfigure()
-        tracer._iast_enabled = True
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/taint-checking-enabled/?q=aaa",
-            headers={"HTTP_USER_AGENT": "test/1.2.3"},
-        )
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_disabled(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)):
-        oce.reconfigure()
-        clear_taint_mapping()
-        setup(bytes.join, bytearray.join)
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/taint-checking-disabled/?q=aaa",
-            headers={"HTTP_USER_AGENT": "test/1.2.3"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled_sqli_http_request_parameter(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/sqli_http_request_parameter/?q=SELECT 1 FROM sqlite_master",
-            headers={"HTTP_USER_AGENT": "test/1.2.3"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [
-            {"origin": "http.request.parameter", "name": "q", "value": "SELECT 1 FROM sqlite_master"}
-        ]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 2588970144
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 FROM sqlite_master", "source": 0}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 79
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled_sqli_http_request_header_value(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/sqli_http_request_header_value/",
-            headers={"HTTP_USER_AGENT": "master"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [{"origin": "http.request.header", "name": "HTTP_USER_AGENT", "value": "master"}]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 1880217241
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 FROM sqlite_"}, {"source": 0, "value": "master"}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 97
-
-        assert response.status_code == 200
-        assert response.content == b"master"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_disabled_sqli_http_request_header_value(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=False
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/sqli_http_request_header_value/",
-            headers={"HTTP_USER_AGENT": "master"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"master"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled_sqli_http_request_header_name(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/sqli_http_request_header_name/",
-            headers={"master": "test/1.2.3"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [{"origin": "http.request.header.name", "name": "master", "value": "master"}]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 3287414465
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 FROM sqlite_"}, {"source": 0, "value": "master"}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 88
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_disabled_sqli_http_request_header_name(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            payload=urlencode({"mytestingbody_key": "mytestingbody_value"}),
-            content_type="application/x-www-form-urlencoded",
-            url="/appsec/sqli_http_request_header_name/",
-            headers={"master": "test/1.2.3"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_iast_enabled_full_sqli_http_path_parameter(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_path_parameter/sqlite_master/",
-            headers={"HTTP_USER_AGENT": "test/1.2.3"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [
-            {"origin": "http.request.path.parameter", "name": "q_http_path_parameter", "value": "sqlite_master"}
-        ]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 2718740693
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 from "}, {"value": "sqlite_master", "source": 0}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 107
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_iast_disabled_full_sqli_http_path_parameter(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=False
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_path_parameter/sqlite_master/",
-            headers={"HTTP_USER_AGENT": "test/1.2.3"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled_sqli_http_cookies_name(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_request_cookie_name/",
-            cookies={"master": "test/1.2.3"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [{"origin": "http.request.cookie.name", "name": "master", "value": "master"}]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 982818537
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 FROM sqlite_"}, {"source": 0, "value": "master"}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 170
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_iast_disabled_sqli_http_cookies_name(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=False
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_request_cookie_name/",
-            cookies={"master": "test/1.2.3"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"test/1.2.3"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_user_agent_iast_enabled_sqli_http_cookies_value(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=True)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=True
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_request_cookie_value/",
-            cookies={"master": "master"},
-        )
-
-        loaded = json.loads(root_span.get_tag(IAST.JSON))
-        assert loaded["sources"] == [{"origin": "http.request.cookie.value", "name": "master", "value": "master"}]
-        assert loaded["vulnerabilities"][0]["type"] == "SQL_INJECTION"
-        assert loaded["vulnerabilities"][0]["hash"] == 893756774
-        assert loaded["vulnerabilities"][0]["evidence"] == {
-            "valueParts": [{"value": "SELECT 1 FROM sqlite_"}, {"source": 0, "value": "master"}]
-        }
-        assert loaded["vulnerabilities"][0]["location"]["line"] == 179
-        assert loaded["vulnerabilities"][0]["location"]["path"] == "tests/contrib/django/django_app/appsec_urls.py"
-
-        assert response.status_code == 200
-        assert response.content == b"master"
-
-
-@pytest.mark.django_db()
-@pytest.mark.skipif(not python_supported_by_iast(), reason="Python version not supported by IAST")
-def test_django_tainted_iast_disabled_sqli_http_cookies_value(client, test_spans, tracer):
-    from ddtrace.appsec.iast._taint_dict import clear_taint_mapping
-    from ddtrace.appsec.iast._taint_tracking import setup
-
-    with override_global_config(dict(_iast_enabled=False)), mock.patch(
-        "ddtrace.contrib.dbapi._is_iast_enabled", return_value=False
-    ):
-        setup(bytes.join, bytearray.join)
-        clear_taint_mapping()
-
-        root_span, response = _aux_appsec_get_root_span(
-            client,
-            test_spans,
-            tracer,
-            url="/appsec/sqli_http_request_cookie_value/",
-            cookies={"master": "master"},
-        )
-
-        assert root_span.get_tag(IAST.JSON) is None
-
-        assert response.status_code == 200
-        assert response.content == b"master"
